+++ src/features/dashboard/views/DashboardPage.vue
@@ -230,7 +230,7 @@
             </article>
           </div>
         </div>
-        <DashboardTaskList />
+        <TaskList />
       </div>
     </div>
   </BoxLoader>
@@ -254,7 +254,7 @@ import { useQuestionnaires } from "../api/useQuestionnaires";
 import { useChangedRecords } from "../api/useChangedRecords";
 import { useArticles } from "../api/useArticles";
 import { useMembers } from "../api/useMembers";
-import DashboardTaskList from "../components/DashboardTaskList.vue";
+import TaskList from "../components/TaskList.vue";
 
 const userStore = useUserStore();
 const { user } = storeToRefs(userStore);
